UncivGame is a pure class file again, GUI split off
